create first version
Bot can now respond to inline queries by showing a list of dice that it supports.
Upon selecting one, a user will be presented with a message showing the number
of sides on the die, the number that was rolled, and, if applicable, the number
of times it was re-rolled. Each message has a re-roll button that can be pressed
to roll the die again.

Upon private-messaging the bot, it will simply send a D20 roll.