move reroll count to button
Decided to move the reroll count to the button instead of having it
at the bottom of the message in favor of visibility. In its original
state the roll would be in the middle of the number of sides on the
die and the reroll count and it would sort of get lost.